Move /broadcast from general_commands to privileged_chatter
