+++ plugins/privileged_chatter.py
@@ -12,17 +12,23 @@ Original authors: kharidiron
 """
 
 from base_plugin import SimpleCommandPlugin
-from plugins.player_manager import Moderator
-from utilities import get_syntax, Command, send_message, ChatReceiveMode
+from plugins.player_manager import Admin, Moderator
+from utilities import get_syntax, Command, send_message, ChatReceiveMode, broadcast
 
 class ModeratorChat(Moderator):
     pass
+
+
+class Broadcast(Admin):
+    pass
+
 
 class PrivilegedChatter(SimpleCommandPlugin):
     name = "privileged_chatter"
     depends = ["command_dispatcher", "chat_enhancements", "player_manager"]
     default_config = {"modchat_color": "^violet;",
-                      "report_prefix": "^magenta;(REPORT): "}
+                      "report_prefix": "^magenta;(REPORT): ",
+                      "broadcast_prefix": "^red;(ADMIN): "}
 
     def __init__(self):
         super().__init__()
@@ -36,6 +42,7 @@ class PrivilegedChatter(SimpleCommandPlugin):
         self.commands = cd.commands
         self.modchat_color = self.config.get_plugin_config(self.name)["modchat_color"]
         self.report_prefix = self.config.get_plugin_config(self.name)["report_prefix"]
+        self.broadcast_prefix = self.config.get_plugin_config(self.name)["broadcast_prefix"]
 
     # Commands - In-game actions that can be performed
 
@@ -66,7 +73,7 @@ class PrivilegedChatter(SimpleCommandPlugin):
 
     @Command("report",
           doc="Privately make a report to all online moderators.")
-    def _moderatorchat(self, data, connection):
+    def _report(self, data, connection):
          """
          Command to send reports to moderators.
 
@@ -87,3 +94,22 @@ class PrivilegedChatter(SimpleCommandPlugin):
                                              name=sender,
                                              mode=send_mode,
                                              channel=channel)
+
+
+    @Command("broadcast",
+         role=Broadcast,
+         doc="Sends a message to everyone on the server.")
+    def _broadcast(self, data, connection):
+        """
+        Broadcast a message to everyone on the server. Currently, this is
+        actually redundant, as sending a message regularly is already a
+        broadcast.
+
+        :param data: The packet containing the command.
+        :param connection: The connection from which the packet came.
+        :return: Null.
+        """
+        if data:
+            message = " ".join(data)
+            broadcast(self,
+                  "{}{}^reset;".format(self.broadcast_prefix, message))
